Move commands into their own files
